move input utils to new file
