+++ typescript/cli/src/utils/chains.ts
@@ -1,13 +1,12 @@
-import { Separator, checkbox, confirm, input } from '@inquirer/prompts';
+import { Separator, checkbox } from '@inquirer/prompts';
 import select from '@inquirer/select';
 import chalk from 'chalk';
 
 import { ChainMap, ChainMetadata } from '@hyperlane-xyz/sdk';
 
-import { log, logGray, logRed, logTip } from '../logger.js';
+import { log, logRed, logTip } from '../logger.js';
 
 import { calculatePageSize } from './cli-options.js';
-import { indentYamlOrJson } from './files.js';
 
 // A special value marker to indicate user selected
 // a new chain in the list
@@ -75,54 +74,4 @@ function handleNewChain(chainNames: string[]) {
     );
     process.exit(0);
   }
-}
-
-export async function detectAndConfirmOrPrompt(
-  detect: () => Promise<string | undefined>,
-  prompt: string,
-  label: string,
-  source?: string,
-): Promise<string> {
-  let detectedValue: string | undefined;
-  try {
-    detectedValue = await detect();
-    if (detectedValue) {
-      const confirmed = await confirm({
-        message: `Detected ${label} as ${detectedValue}${
-          source ? ` from ${source}` : ''
-        }, is this correct?`,
-      });
-      if (confirmed) {
-        return detectedValue;
-      }
-    }
-    // eslint-disable-next-line no-empty
-  } catch (e) {}
-  return input({ message: `${prompt} ${label}:`, default: detectedValue });
-}
-
-const INFO_COMMAND: string = 'i';
-const DOCS_NOTICE: string =
-  'For more information, please visit https://docs.hyperlane.xyz.';
-
-export async function inputWithInfo({
-  message,
-  info = 'No additional information available.',
-  defaultAnswer,
-}: {
-  message: string;
-  info?: string;
-  defaultAnswer?: string;
-}): Promise<string> {
-  let answer: string = '';
-  do {
-    answer = await input({
-      message: message.concat(` [enter '${INFO_COMMAND}' for more info]`),
-      default: defaultAnswer,
-    });
-    answer = answer.trim().toLowerCase();
-    const indentedInfo = indentYamlOrJson(`${info}\n${DOCS_NOTICE}\n`, 4);
-    if (answer === INFO_COMMAND) logGray(indentedInfo);
-  } while (answer === INFO_COMMAND);
-  return answer;
 }
